requirements.txt
asttokens==2.4.1
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
Flask==3.1.0
idna==3.10
ipykernel==6.29.4
ipython==8.25.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nodeenv==1.9.1
packaging==24.1
parso==0.8.4
pathspec==0.12.1
platformdirs==4.3.8
prompt_toolkit==3.0.47
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.18.0
pyright==1.1.400
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pywin32==306
pyzmq==26.0.3
requests==2.32.3
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
